+++ database/migrations/2019_06_24_183340_create_products_table.php
@@ -19,7 +19,7 @@ class CreateProductsTable extends Migration
             $table->string('slug');
             $table->text('description');
             $table->decimal('price');
-            $table->unsignedInteger('stock')->default(0);
+            $table->decimal('stock')->default(0);
             $table->boolean('active')->default(1);
             $table->timestamps();
         });
